requirements.txt
beautifulsoup4
requests
retrying
pysocks
flask
python-telegram-bot
python-dotenv